@@ -37,7 +37,7 @@ project.ext.versionCode = 1_000_000 * major + 10_000 * minor + 100 * patch + bui
 android {
     flavorDimensions "testapp"
 
-    compileSdkVersion 27
+    compileSdkVersion 28
     signingConfigs {
         release {
             keyAlias 'EDRLab'
