@@ -10,7 +10,7 @@
 
 package org.readium.r2.testapp
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 
 class DRMManagementActivity : AppCompatActivity() {
