@@ -4,18 +4,20 @@ import android.app.ProgressDialog
 import android.media.MediaPlayer
 import android.media.MediaPlayer.OnPreparedListener
 import android.net.Uri
-import android.os.Handler
-import org.readium.r2.navigator.pager.PageCallback
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.readium.r2.shared.Link
-
 import java.io.IOException
 
 class R2MediaPlayer(var mediaActivity: AudiobookActivity, var items: MutableList<Link>, var progress: ProgressDialog, var callback: MediaPlayerCallback) : OnPreparedListener {
 
+    private val uiScope = CoroutineScope(Dispatchers.Main)
+
     var mediaPlayer: MediaPlayer = MediaPlayer()
 
     val isPlaying: Boolean
-        get() = mediaPlayer.isPlaying()
+        get() = mediaPlayer.isPlaying
 
     val duration: Double
         get() = mediaPlayer.duration.toDouble()
@@ -67,9 +69,9 @@ class R2MediaPlayer(var mediaActivity: AudiobookActivity, var items: MutableList
     }
 
     fun toggleProgress(show: Boolean) {
-        mediaActivity.runOnUiThread {
-            if (show) progress.show();
-            else progress.hide();
+        uiScope.launch {
+            if (show) progress.show()
+            else progress.hide()
         }
     }
 
