@@ -11,10 +11,10 @@
 package org.readium.r2.testapp.opds
 
 import android.content.Context
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.util.TypedValue
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 
 class GridAutoFitLayoutManager : GridLayoutManager {
     private var mColumnWidth: Int = 0
