@@ -11,11 +11,8 @@
 package org.readium.r2.testapp.opds
 
 import android.app.Activity
+import android.content.Context
 import android.os.Bundle
-import android.support.v7.app.AlertDialog
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.text.TextUtils
 import android.view.Gravity
 import android.view.View
@@ -24,6 +21,9 @@ import android.webkit.URLUtil
 import android.widget.Button
 import android.widget.EditText
 import android.widget.LinearLayout
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
 import com.github.kittinunf.fuel.Fuel
 import com.mcxiaoke.koi.ext.onClick
 import com.mcxiaoke.koi.ext.onLongClick
@@ -54,11 +54,29 @@ class OPDSListActivity : AppCompatActivity() {
 
         val database = OPDSDatabase(this)
 
-        database.opds.insert(OPDSModel("Feedbooks", "http://www.feedbooks.com/catalog.atom", 1))
-        database.opds.insert(OPDSModel("Open Textbooks", "http://open.minitex.org/", 1))
+        val preferences = getSharedPreferences("org.readium.r2.testapp", Context.MODE_PRIVATE)
+
+        val version = 2
+        val VERSION_KEY = "OPDS_CATALOG_VERSION"
+
+        if (preferences.getInt(VERSION_KEY, 0) < version) {
+            preferences.edit().putInt(VERSION_KEY, version).apply()
+
+            database.opds.emptyTable()
+
+            val R2TestCatalog = OPDSModel( "R2 Reader Test Catalog",  "https://d2g.dita.digital/opds/collections/10040", 1)
+            val OPDS2Catalog = OPDSModel( "OPDS 2.0 Test Catalog",  "https://test.opds.io/2.0/home.json", 2)
+            val OTBCatalog = OPDSModel( "Open Textbooks Catalog",  "http://open.minitex.org/textbooks/", 1)
+            val SEBCatalog = OPDSModel( "Standard eBooks Catalog",  " https://standardebooks.org/opds/all", 1)
+
+            database.opds.insert(R2TestCatalog)
+            database.opds.insert(OPDS2Catalog)
+            database.opds.insert(OTBCatalog)
+            database.opds.insert(SEBCatalog)
+        }
 
         val list = database.opds.list().toMutableList()
-        val opdsAdapter = OPDSViewAdapter(act, list)
+        val opdsAdapter = OPDSViewAdapter(this, list)
 
         coordinatorLayout {
             fitsSystemWindows = true
@@ -70,8 +88,8 @@ class OPDSListActivity : AppCompatActivity() {
                 linearLayout {
                     orientation = LinearLayout.VERTICAL
                     recyclerView {
-                        layoutManager = LinearLayoutManager(act)
-                        (layoutManager as LinearLayoutManager).orientation = LinearLayoutManager.VERTICAL
+                        layoutManager = androidx.recyclerview.widget.LinearLayoutManager(this@OPDSListActivity)
+                        (layoutManager as androidx.recyclerview.widget.LinearLayoutManager).orientation = RecyclerView.VERTICAL
                         adapter = opdsAdapter
                     }
                 }
@@ -170,7 +188,7 @@ class OPDSListActivity : AppCompatActivity() {
     }
 }
 
-private class OPDSViewAdapter(private val activity: Activity, private val list: MutableList<OPDSModel>) : RecyclerView.Adapter<OPDSViewAdapter.ViewHolder>() {
+private class OPDSViewAdapter(private val activity: Activity, private val list: MutableList<OPDSModel>) : androidx.recyclerview.widget.RecyclerView.Adapter<OPDSViewAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val inflater = activity.layoutInflater
@@ -203,7 +221,7 @@ private class OPDSViewAdapter(private val activity: Activity, private val list: 
         return list.size
     }
 
-    internal inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+    internal inner class ViewHolder(view: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(view) {
         val button: Button = view.findViewById<View>(R.id.button) as Button
 
     }
